Client1/requirements.txt
customtkinter==5.2.1
opencv-python==4.8.1.78
mediapipe==0.10.8
pillow==10.1.0
requests==2.31.0
numpy==1.26.2
joblib==1.3.2
scikit-learn
pandas
